rename mapnik-cli to pymapnik-cli
Mapnik is the system module name. If we name our file the same we try to
import ourself and not the system mapnik module. Moreover add main entry
so that we can run this command standalone. This is needed because the
flask app won't be instantiable if mapnik is not installed (chicken-egg
problem)